EVG-16570 Fix lint
