+++ model/task/task_testresult_test.go
@@ -1,7 +1,9 @@
 package task
 
 import (
+	"context"
 	"fmt"
+	"sort"
 	"testing"
 	"time"
 
@@ -9,6 +11,8 @@ import (
 	"github.com/evergreen-ci/evergreen/db"
 	mgobson "github.com/evergreen-ci/evergreen/db/mgo/bson"
 	"github.com/evergreen-ci/evergreen/model/testresult"
+	"github.com/evergreen-ci/evergreen/testutil"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
 )
 
@@ -193,4 +197,97 @@ func (s *TaskTestResultSuite) TestArchivedTask() {
 	s.Equal("revision-20", t.Revision)
 
 	s.Len(t.LocalTestResults, 9)
+}
+
+func TestGetTestCountByTaskIdAndFilter(t *testing.T) {
+	assert := assert.New(t)
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	env := testutil.NewEnvironment(ctx, t)
+	evergreen.SetEnvironment(env)
+	assert.NoError(db.ClearCollections(Collection, testresult.Collection))
+
+	numTests := 10
+	numTasks := 2
+	testObjects := make([]string, numTests)
+
+	for ix := range testObjects {
+		testObjects[ix] = fmt.Sprintf("TestSuite/TestNum%d", ix)
+	}
+	sort.StringSlice(testObjects).Sort()
+	for i := 0; i < numTasks; i++ {
+		id := fmt.Sprintf("task_%d", i)
+		testTask := &Task{
+			Id: id,
+		}
+		tests := make([]TestResult, numTests)
+		for j := 0; j < numTests; j++ {
+			status := "pass"
+			if j%2 == 0 {
+				status = "fail"
+			}
+			tests[j] = TestResult{
+				TaskID:    id,
+				Execution: 0,
+				Status:    status,
+				TestFile:  testObjects[j],
+				EndTime:   float64(j),
+				StartTime: 0,
+			}
+		}
+		assert.NoError(testTask.Insert())
+		for _, test := range tests {
+			assert.NoError(test.Insert())
+		}
+	}
+
+	for i := 0; i < numTasks; i++ {
+		taskId := fmt.Sprintf("task_%d", i)
+		count, err := GetTestCountByTaskIdAndFilters(taskId, "", []string{}, 0)
+		assert.NoError(err)
+		assert.Equal(count, numTests)
+
+		count, err = GetTestCountByTaskIdAndFilters(taskId, "", []string{"pass"}, 0)
+		assert.NoError(err)
+		assert.Equal(count, numTests/2)
+
+		count, err = GetTestCountByTaskIdAndFilters(taskId, "", []string{"fail"}, 0)
+		assert.NoError(err)
+		assert.Equal(count, numTests/2)
+
+		count, err = GetTestCountByTaskIdAndFilters(taskId, "", []string{"pass", "fail"}, 0)
+		assert.NoError(err)
+		assert.Equal(count, 10)
+
+		count, err = GetTestCountByTaskIdAndFilters(taskId, "TestSuite/TestNum1", []string{}, 0)
+		assert.NoError(err)
+		assert.Equal(count, 1)
+
+		count, err = GetTestCountByTaskIdAndFilters(taskId, "TestSuite/TestNum2", []string{}, 0)
+		assert.NoError(err)
+		assert.Equal(count, 1)
+
+		count, err = GetTestCountByTaskIdAndFilters(taskId, "TestSuite/TestN", []string{}, 0)
+		assert.NoError(err)
+		assert.Equal(count, numTests)
+
+		count, err = GetTestCountByTaskIdAndFilters(taskId, "TestSuite/TestN", []string{"pass", "fail"}, 0)
+		assert.NoError(err)
+		assert.Equal(count, numTests)
+
+		count, err = GetTestCountByTaskIdAndFilters(taskId, "TestSuite/TestN", []string{"pass"}, 0)
+		assert.NoError(err)
+		assert.Equal(count, numTests/2)
+
+		count, err = GetTestCountByTaskIdAndFilters(taskId, "", []string{"pa"}, 0)
+		assert.NoError(err)
+		assert.Equal(count, 0)
+
+		count, err = GetTestCountByTaskIdAndFilters(taskId, "", []string{"not_a_real_status"}, 0)
+		assert.NoError(err)
+		assert.Equal(count, 0)
+	}
+	count, err := GetTestCountByTaskIdAndFilters("fake_task", "", []string{}, 0)
+	assert.Error(err)
+	assert.Equal(count, 0)
 }
